Reorganized rotation files for frost, sub unholy still broken
