--- sim/deathknight/dps/rotation_frost_helper.go
@@ -1,11 +1,0 @@
-package dps
-
-import (
-	"github.com/wowsims/wotlk/sim/core"
-)
-
-type FrostRotation struct {
-}
-
-func (fr *FrostRotation) Reset(sim *core.Simulation) {
-}
