+++ sim/deathknight/dps/rotation_frost_sub_unholy.go
@@ -1,0 +1,195 @@
+package dps
+
+import (
+	"time"
+
+	"github.com/wowsims/wotlk/sim/core"
+	"github.com/wowsims/wotlk/sim/deathknight"
+)
+
+func (dk *DpsDeathknight) setupFrostSubUnholyERWOpener() {
+	dk.Opener.
+		NewAction(dk.RotationActionCallback_IT).
+		NewAction(dk.RotationActionCallback_PS).
+		NewAction(dk.RotationActionCallback_BT).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Pesti).
+		NewAction(dk.RotationActionCallback_UA).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_FS_Star).
+		NewAction(dk.RotationActionCallback_ERW).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FS).
+		NewAction(dk.RotationActionCallback_FS).
+		NewAction(dk.RotationActionCallback_FS).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_BS).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Pesti).
+		NewAction(dk.RotationActionCallback_FS)
+
+	dk.Main.
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Step1)
+}
+
+func (dk *DpsDeathknight) setupFrostSubUnholyNoERWOpener() {
+	dk.Opener.
+		NewAction(dk.RotationActionCallback_IT).
+		NewAction(dk.RotationActionCallback_PS).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_BS).
+		NewAction(dk.RotationActionCallback_BT).
+		NewAction(dk.RotationActionCallback_UA).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Pesti).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_FS_Star).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FS).
+		NewAction(dk.RotationActionCallback_FS).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_BS).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Pesti).
+		NewAction(dk.RotationActionCallback_FS)
+
+	dk.Main.
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Step1)
+}
+
+func (dk *DpsDeathknight) RotationActionCallback_FrostSubUnholy_Step1(sim *core.Simulation, target *core.Unit, s *deathknight.Sequence) bool {
+	s.Clear().
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Pesti).
+		NewAction(dk.RotationActionCallback_BS).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_FS_Star).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_FS_Star).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Step2)
+	return false
+}
+
+func (dk *DpsDeathknight) RotationActionCallback_FrostSubUnholy_Step2(sim *core.Simulation, target *core.Unit, s *deathknight.Sequence) bool {
+	s.Clear().
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FS).
+		NewAction(dk.RotationActionCallback_UA).
+		NewAction(dk.RotationActionCallback_BT).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_FS_Star).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_FS_Star).
+		NewAction(dk.RotationActionCallback_FrostSubUnholy_Step1)
+	return false
+}
+
+func (dk *DpsDeathknight) RotationActionCallback_FrostSubUnholy_RecoverFromPestiMiss(sim *core.Simulation, target *core.Unit, s *deathknight.Sequence) bool {
+	if dk.LastCast == dk.BloodStrike {
+		s.Clear().
+			NewAction(dk.RotationActionCallback_BS).
+			NewAction(dk.RotationActionCallback_FS).
+			NewAction(dk.RotationActionCallback_IT).
+			NewAction(dk.RotationActionCallback_PS).
+			NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+			NewAction(dk.RotationActionCallback_BS).
+			NewAction(dk.RotationActionCallback_FrostSubUnholy_Pesti).
+			NewAction(dk.RotationActionCallback_FrostSubUnholy_Step1)
+	} else {
+		s.Clear().
+			NewAction(dk.RotationActionCallback_BS).
+			NewAction(dk.RotationActionCallback_BS).
+			NewAction(dk.RotationActionCallback_FS).
+			NewAction(dk.RotationActionCallback_IT).
+			NewAction(dk.RotationActionCallback_PS).
+			NewAction(dk.RotationActionCallback_FrostSubUnholy_Obli).
+			NewAction(dk.RotationActionCallback_BS).
+			NewAction(dk.RotationActionCallback_FrostSubUnholy_Pesti).
+			NewAction(dk.RotationActionCallback_FrostSubUnholy_Step1)
+	}
+
+	return false
+}
+
+func (dk *DpsDeathknight) RotationActionCallback_FrostSubUnholy_Obli(sim *core.Simulation, target *core.Unit, s *deathknight.Sequence) bool {
+	casted := false
+	ffActive := dk.FrostFeverDisease[target.Index].IsActive()
+	bpActive := dk.BloodPlagueDisease[target.Index].IsActive()
+
+	if ffActive && bpActive {
+		ffExpiresAt := dk.FrostFeverDisease[target.Index].ExpiresAt()
+		bpExpiresAt := dk.BloodPlagueDisease[target.Index].ExpiresAt()
+
+		if dk.CanObliterate(sim) {
+			runeCost := dk.OptimalRuneCost(core.RuneCost(dk.Obliterate.DefaultCast.Cost))
+
+			if runeCost.Death() == 0 {
+				casted = dk.CastObliterate(sim, target)
+				s.ConditionalAdvance(casted)
+			} else {
+				if sim.CurrentTime+(10*time.Second-dk.CurrentBloodRuneGrace(sim)) < core.MinDuration(ffExpiresAt, bpExpiresAt) {
+					casted = dk.CastObliterate(sim, target)
+					s.ConditionalAdvance(casted)
+				} else {
+					s.Advance()
+				}
+
+			}
+		}
+	} else {
+		s.Advance()
+	}
+
+	return casted
+}
+
+func (dk *DpsDeathknight) RotationActionCallback_FrostSubUnholy_Pesti(sim *core.Simulation, target *core.Unit, s *deathknight.Sequence) bool {
+	casted := false
+
+	ffActive := dk.FrostFeverDisease[target.Index].IsActive()
+	bpActive := dk.BloodPlagueDisease[target.Index].IsActive()
+
+	if !ffActive || !bpActive {
+		dk.RotationActionCallback_FrostSubUnholy_RecoverFromPestiMiss(sim, target, s)
+		return casted
+	} else {
+		casted = dk.CastPestilence(sim, target)
+		advance := dk.LastOutcome.Matches(core.OutcomeLanded)
+		if !casted || (casted && !dk.LastOutcome.Matches(core.OutcomeLanded)) {
+			ffExpiresAt := dk.FrostFeverDisease[target.Index].ExpiresAt()
+			bpExpiresAt := dk.BloodPlagueDisease[target.Index].ExpiresAt()
+
+			if sim.CurrentTime+dk.SpellGCD() > ffExpiresAt || sim.CurrentTime+dk.SpellGCD() > bpExpiresAt {
+				dk.RotationActionCallback_FrostSubUnholy_RecoverFromPestiMiss(sim, target, s)
+				return casted
+			} else {
+				s.ConditionalAdvance(casted && advance)
+				return casted
+			}
+		} else {
+			s.ConditionalAdvance(casted && advance)
+			return casted
+		}
+	}
+
+	return casted
+}
+
+func (dk *DpsDeathknight) RotationActionCallback_FrostSubUnholy_FS_Star(sim *core.Simulation, target *core.Unit, s *deathknight.Sequence) bool {
+	casted := false
+	if dk.PercentRunicPower() >= 0.95 || (dk.KillingMachineAura.IsActive() && dk.CurrentRunicPower() >= 32.0) {
+		casted = dk.CastFrostStrike(sim, target)
+	} else if dk.RimeAura.IsActive() {
+		casted = dk.CastHowlingBlast(sim, target)
+	} else if dk.CurrentRunicPower() >= 32.0 {
+		casted = dk.CastFrostStrike(sim, target)
+		if !casted {
+			casted = dk.CastHornOfWinter(sim, target)
+		}
+	} else {
+		casted = false
+	}
+
+	s.Advance()
+	return casted
+}
